Split grammarMatcher.spec.ts into themed test files
Split the monolithic 1984-line grammarMatcher.spec.ts into 7 focused
test files organized by theme:

- grammarMatcherBasic.spec.ts: basic values, match, escaped match, not-matched
- grammarMatcherVariables.spec.ts: variable types, nested rules, wildcards
- grammarMatcherGroups.spec.ts: repeat groups, recursive rules
- grammarMatcherPunctuation.spec.ts: punctuation as separator
- grammarMatcherSpacingBasic.spec.ts: core spacing modes, CJK/Hangul/digit
  boundaries, parse errors, number variable spacing, repeat groups with modes
- grammarMatcherSpacingNone.spec.ts: all spacing=none specific tests
- grammarMatcherSpacingNested.spec.ts: mode switching, merged rules, deeply
  nested separator behavior

Moved the shared testMatchGrammar helper into testUtils.ts.
No test logic was added or changed.
